+++ src/classes/ProviderError.ts
@@ -19,9 +19,11 @@ export const ProviderErrors = {
   PairingFailed: new ProviderError('Pairing failed', 4900),
   Disconnected: new ProviderError('Disconnected', 4900),
   UnsupportedMethod: new ProviderError('Unsupported Method', 4200),
-  Unauthorized: new ProviderError('Unauthorized', 4900),
+  Unauthorized: new ProviderError('Unauthorized', 4100),
   NoResult: new ProviderError('No result', 4900),
   NoAppId: new ProviderError('App ID is required', 4900),
+  UserRejected: new ProviderError('User Rejected Request', 4001),
+  ChainDisconnected: new ProviderError('Chain Disconnected', 4901),
 };
 
 export const newProviderError = (error?: unknown): ProviderError => {