+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Grindery Wallet SDK Changelog
 
+## v0.3.0 (2024-06-18)
+
+### Added support for BNB and opBNB chains
+
+- Added `eth_chainId` provider method, as specified by [EIP-695](https://eips.ethereum.org/EIPS/eip-695)
+- Added `wallet_addEthereumChain` provider method, as specified by [EIP-3085](https://eips.ethereum.org/EIPS/eip-3085)
+- Added `wallet_switchEthereumChain` provider method, as specified by [EIP-3326](https://eips.ethereum.org/EIPS/eip-3326)
+- Added `chainChanged` provider event, as specified by [EIP-1193](https://eips.ethereum.org/EIPS/eip-1193)
+
 ## v0.2.0 (2024-05-16)
 
 ### Improvements